src/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.6.1
distro==1.9.0
filelock==3.16.1
flake8==7.1.2
Flake8-pyproject==1.2.3
fsspec==2024.10.0
GitPython==3.1.43
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
jiter==0.9.1
loguru==0.7.3
mccabe==0.7.0
openai==1.107.2
packaging==25.0
pluggy==1.5.0
pycodestyle==2.12.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.2
pytest==8.3.5
PyYAML==6.0.2
requests==2.32.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
mangum
fastapi
uvicorn
PyJWT
python-multipart
types-PyYAML
